Added &find_files to run_cmd and minor tweaks.

+++ bwa.pm
@@ -15,20 +15,20 @@ Title   : bwa_align
 Usage   : my $mapped_bam = bwa_aling($input,$reference,{other_options => possible})
 Function: BWA align an input file against a reference.
 Args    : 
-	input=			*Mandatory* Input file to be BWA mapped. Either: in.bam or in_1.fq,in_2.fq
-	ref=			*Mandatory* Reference.fna+index
-	{
-		output_prefix	=>	Prefix for each output.  Ie. (SRA_LGT)_at_\$ref_name
-		output_dir		=>	Will output to current working directory unless another is specified with this. ie. /ksieber_dir/tmp/
-		sort_index 		=>	<0|1> [0] 1= Sort and index the new.bam into new.srt.bam and new.srt.bai. 
-		M_only 			=>	<0|1> [0] 1= Only keep mates with M_*
-		MM_only			=>	<0|1> [0] 1= Only keep mates with M_M
-		mpileup 		=>	<0|1> [0] 1= Calculate pileup coverage on .bam.
-		insert_metrics 	=> 	<0|1> [0] 1= Use Picard to calculate insert size metrics.
-		disable_SW		=>	<0|1> [0] 1= Disable Smith-Waterman for the UM mate. Ideal for quicker LGT mappings IF they are high confidence. 
-		sam_output		=>	<0|1> [0] 0= .bam output; 1= .sam output
-		no_cleanup 		=>	<0|1> [0] 0= Removes .sai files and unsorted.bam with --sort_index. 1=No deleting intermediate data. 
-	}
+    input=          *Mandatory* Input file to be BWA mapped. Either: in.bam or in_1.fq,in_2.fq
+    ref=            *Mandatory* Reference.fna+index
+    {
+        output_prefix   =>  Prefix for each output.  Ie. (SRA_LGT)_at_\$ref_name
+        output_dir      =>  Will output to current working directory unless another is specified with this. ie. /ksieber_dir/tmp/
+        sort_index      =>  <0|1> [0] 1= Sort and index the new.bam into new.srt.bam and new.srt.bai. 
+        M_only          =>  <0|1> [0] 1= Only keep mates with M_*
+        MM_only         =>  <0|1> [0] 1= Only keep mates with M_M
+        mpileup         =>  <0|1> [0] 1= Calculate pileup coverage on .bam.
+        insert_metrics  =>  <0|1> [0] 1= Use Picard to calculate insert size metrics.
+        disable_SW      =>  <0|1> [0] 1= Disable Smith-Waterman for the UM mate. Ideal for quicker LGT mappings IF they are high confidence. 
+        sam_output      =>  <0|1> [0] 0= .bam output; 1= .sam output
+        no_cleanup      =>  <0|1> [0] 0= Removes .sai files and unsorted.bam with --sort_index. 1=No deleting intermediate data. 
+    }
 
 Returns : File path to a newly mapped Bam.
 
@@ -41,11 +41,10 @@ sub bwa_align {
     my @ref_suffix_list = ( '.fasta', '.fa', '.fna', '.txt' );
     my $threads = defined $options->{t} ? "$options->{t}" : "1";
 
-    my ( $ref_name, $ref_path, $ref_suf )
-        = fileparse( $ref, @ref_suffix_list );    ## Grab the reference name to use for the naming the output
-    my $file1;                                    ## Global input file name
-    my $file2;                                    ## Global input file name2
-    my $bam;                                      ## 0=fastq, 1=bam
+    my ( $ref_name, $ref_path, $ref_suf ) = fileparse( $ref, @ref_suffix_list );    ## Grab the reference name to use for the naming the output
+    my $file1;                                                                      ## Global input file name
+    my $file2;                                                                      ## Global input file name2
+    my $bam;                                                                        ## 0=fastq, 1=bam
     if ( $files =~ /.fq$/ || $files =~ /.fastq$/ || $files =~ /.fastq.gz$/ ) {
         ( $file1, $file2 ) = split( /,/, $files );
         if ( $file2 !~ /\w+/ ) { $file2 = $file1; }
@@ -75,83 +74,59 @@ sub bwa_align {
 
     ## Run BWA ALN
     if ( $bam == 1 ) {
-        run_cmd(
-            "bwa aln -e -1 -M 3 -E 4 -O 11 -t $threads -o 1 $ref -b1 $file1 > $output_prefix\.1.sai 2>>$output_prefix\_bwa_stderr.log",
-            $log
-        );
-        run_cmd(
-            "bwa aln -e -1 -M 3 -E 4 -O 11 -t $threads -o 1 $ref -b2 $file2 > $output_prefix\.2.sai 2>>$output_prefix\_bwa_stderr.log",
-            $log
-        );
+        run_cmd( "bwa aln -e -1 -M 3 -E 4 -O 11 -t $threads -o 1 $ref -b1 $file1 > $output_prefix\.1.sai 2>>$output_prefix\_bwa_stderr.log", $log );
+        run_cmd( "bwa aln -e -1 -M 3 -E 4 -O 11 -t $threads -o 1 $ref -b2 $file2 > $output_prefix\.2.sai 2>>$output_prefix\_bwa_stderr.log", $log );
     }
     else {
-        run_cmd(
-            "bwa aln -e -1 -M 3 -E 4 -O 11 -t $threads -o 1 $ref $file1 > $output_prefix\.1.sai 2>>$output_prefix\_bwa_stderr.log",
-            $log
-        );
-        run_cmd(
-            "bwa aln -e -1 -M 3 -E 4 -O 11 -t $threads -o 1 $ref $file2 > $output_prefix\.2.sai 2>>$output_prefix\_bwa_stderr.log",
-            $log
-        );
+        run_cmd( "bwa aln -e -1 -M 3 -E 4 -O 11 -t $threads -o 1 $ref $file1 > $output_prefix\.1.sai 2>>$output_prefix\_bwa_stderr.log", $log );
+        run_cmd( "bwa aln -e -1 -M 3 -E 4 -O 11 -t $threads -o 1 $ref $file2 > $output_prefix\.2.sai 2>>$output_prefix\_bwa_stderr.log", $log );
     }
 
     ## BWA SAMPE
     ## Comment in our out according to samtools version
     ## samtools 0.2.0 (Broken samtools sort & new view options)
 # if ($options->{M_only}){
-# 	run_cmd("bwa sampe $ref $output_prefix\.1.sai $output_prefix\.2.sai $file1 $file2 | samtools view -b -F0x4 - > $output_prefix\.bam",$log);
+#   run_cmd("bwa sampe $ref $output_prefix\.1.sai $output_prefix\.2.sai $file1 $file2 | samtools view -b -F0x4 - > $output_prefix\.bam",$log);
 # } elsif ($options->{MM_only}){
-# 	run_cmd("bwa sampe $ref $output_prefix\.1.sai $output_prefix\.2.sai $file1 $file2 | samtools view -b -F0xC - > $output_prefix\.bam",$log);
+#   run_cmd("bwa sampe $ref $output_prefix\.1.sai $output_prefix\.2.sai $file1 $file2 | samtools view -b -F0xC - > $output_prefix\.bam",$log);
 # } elsif ($options->{sam_output}) {
-# 	if ($options->{disable_SW}) {
-# 		run_cmd("bwa sampe -s $ref $output_prefix\.1.sai $output_prefix\.2.sai $file1 $file2 > $output_prefix\.sam 2>>$output_prefix\_bwa_stderr.log",$log);
-# 	} else {
-# 		run_cmd("bwa sampe $ref $output_prefix\.1.sai $output_prefix\.2.sai $file1 $file2 > $output_prefix\.sam 2>>$output_prefix\_bwa_stderr.log",$log);
-# 	}
+#   if ($options->{disable_SW}) {
+#       run_cmd("bwa sampe -s $ref $output_prefix\.1.sai $output_prefix\.2.sai $file1 $file2 > $output_prefix\.sam 2>>$output_prefix\_bwa_stderr.log",$log);
+#   } else {
+#       run_cmd("bwa sampe $ref $output_prefix\.1.sai $output_prefix\.2.sai $file1 $file2 > $output_prefix\.sam 2>>$output_prefix\_bwa_stderr.log",$log);
+#   }
 # } else {
-# 	if ($options->{disable_SW}) {
-# 		run_cmd("bwa sampe -s $ref $output_prefix\.1.sai $output_prefix\.2.sai $file1 $file2 2>>$output_prefix\_bwa_stderr.log | samtools view -b - > $output_prefix\.bam ",$log);
-# 	} else {
-# 		run_cmd("bwa sampe $ref $output_prefix\.1.sai $output_prefix\.2.sai $file1 $file2 2>>$output_prefix\_bwa_stderr.log | samtools view -b - > $output_prefix\.bam ",$log);  ## 2>>$output_prefix\_bwa_stderr.log
-# 	}
+#   if ($options->{disable_SW}) {
+#       run_cmd("bwa sampe -s $ref $output_prefix\.1.sai $output_prefix\.2.sai $file1 $file2 2>>$output_prefix\_bwa_stderr.log | samtools view -b - > $output_prefix\.bam ",$log);
+#   } else {
+#       run_cmd("bwa sampe $ref $output_prefix\.1.sai $output_prefix\.2.sai $file1 $file2 2>>$output_prefix\_bwa_stderr.log | samtools view -b - > $output_prefix\.bam ",$log);  ## 2>>$output_prefix\_bwa_stderr.log
+#   }
 # }
 
     if ( $options->{M_only} ) {
-        run_cmd(
-            "bwa sampe $ref $output_prefix\.1.sai $output_prefix\.2.sai $file1 $file2 | samtools view -bhS -F0x4 - > $output_prefix\.bam",
-            $log
-        );
+        run_cmd( "bwa sampe $ref $output_prefix\.1.sai $output_prefix\.2.sai $file1 $file2 | samtools view -bhS -F0x4 - > $output_prefix\.bam 2>>$output_prefix\_bwa_stderr.log", $log );
     }
     elsif ( $options->{MM_only} ) {
-        run_cmd(
-            "bwa sampe $ref $output_prefix\.1.sai $output_prefix\.2.sai $file1 $file2 | samtools view -bhS -F0xC - > $output_prefix\.bam",
-            $log
-        );
+        run_cmd( "bwa sampe $ref $output_prefix\.1.sai $output_prefix\.2.sai $file1 $file2 | samtools view -bhS -F0xC - > $output_prefix\.bam 2>>$output_prefix\_bwa_stderr.log", $log );
     }
     elsif ( $options->{sam_output} ) {
         if ( $options->{disable_SW} ) {
-            run_cmd(
-                "bwa sampe -s $ref $output_prefix\.1.sai $output_prefix\.2.sai $file1 $file2 > $output_prefix\.sam 2>>$output_prefix\_bwa_stderr.log",
-                $log
-            );
+            run_cmd( "bwa sampe -s $ref $output_prefix\.1.sai $output_prefix\.2.sai $file1 $file2 > $output_prefix\.sam 2>>$output_prefix\_bwa_stderr.log", $log );
         }
         else {
-            run_cmd(
-                "bwa sampe $ref $output_prefix\.1.sai $output_prefix\.2.sai $file1 $file2 > $output_prefix\.sam 2>>$output_prefix\_bwa_stderr.log",
-                $log
-            );
+            run_cmd( "bwa sampe $ref $output_prefix\.1.sai $output_prefix\.2.sai $file1 $file2 > $output_prefix\.sam 2>>$output_prefix\_bwa_stderr.log", $log );
         }
     }
     else {
         if ( $options->{disable_SW} ) {
             run_cmd(
-                "bwa sampe -s $ref $output_prefix\.1.sai $output_prefix\.2.sai $file1 $file2 2>>$output_prefix\_bwa_stderr.log | samtools view -bhS - > $output_prefix\.bam ",
+                "bwa sampe -s $ref $output_prefix\.1.sai $output_prefix\.2.sai $file1 $file2 2>>$output_prefix\_bwa_stderr.log | samtools view -bhS - > $output_prefix\.bam 2>>$output_prefix\_bwa_stderr.log",
                 $log
             );
         }
         else {
             run_cmd(
-                "bwa sampe $ref $output_prefix\.1.sai $output_prefix\.2.sai $file1 $file2 2>>$output_prefix\_bwa_stderr.log | samtools view -bhS - > $output_prefix\.bam ",
+                "bwa sampe $ref $output_prefix\.1.sai $output_prefix\.2.sai $file1 $file2 2>>$output_prefix\_bwa_stderr.log | samtools view -bhS - > $output_prefix\.bam 2>>$output_prefix\_bwa_stderr.log",
                 $log
             );    ## 2>>$output_prefix\_bwa_stderr.log
         }
@@ -161,18 +136,18 @@ sub bwa_align {
     ## Comment in our out according to samtools version
     ## Use this with samtools version >= 0.2.0 (broken)
     # if ($options->{sort_index}){
-    # 	run_cmd("samtools sort -o $output_prefix\.bam - | samtools view -bh - > $output_prefix\.srt.bam",$log);
-    # 	run_cmd("samtools index $output_prefix\.srt.bam",$log);
+    #   run_cmd("samtools sort -o $output_prefix\.bam - | samtools view -bh - > $output_prefix\.srt.bam",$log);
+    #   run_cmd("samtools index $output_prefix\.srt.bam",$log);
     # }
     ## Use this with samtools version <= 0.1.19
     if ( $options->{sort_index} ) {
-        run_cmd( "samtools sort $output_prefix\.bam $output_prefix\.srt", $log );
-        run_cmd( "samtools index $output_prefix\.srt.bam",                $log );
+        run_cmd( "samtools sort $output_prefix\.bam $output_prefix\.srt 2>>$output_prefix\_bwa_stderr.log", $log );
+        run_cmd( "samtools index $output_prefix\.srt.bam 2>>$output_prefix\_bwa_stderr.log",                $log );
     }
 
     ## Samtools Mpileup
     if ( $options->{mpileup} ) {
-        run_cmd( "samtools mpileup -Af $options->{ref} $output_prefix\.srt.bam > $output_prefix\.COVERAGE.txt", $log );
+        run_cmd( "samtools mpileup -Af $options->{ref} $output_prefix\.srt.bam > $output_prefix\.COVERAGE.txt 2>>$output_prefix\_bwa_stderr.log", $log );
     }
 
     ## Picard insert metrics
@@ -188,7 +163,7 @@ sub bwa_align {
     }
 
     ## Cleanup intermediate files (.sai)
-    unless ( $options->{no_cleanup} ) {
+    unless ( defined $options->{no_cleanup} and $options->{no_cleanup} == 1 ) {
         run_cmd( "rm $output_prefix\.1.sai",          $log );
         run_cmd( "rm $output_prefix\.2.sai",          $log );
         run_cmd( "rm $output_prefix\_bwa_stderr.log", $log );
